+++ app/Owner.php
@@ -5,13 +5,14 @@ namespace App;
 use LaravelEnso\Contacts\app\Traits\Contactable;
 use LaravelEnso\Core\app\Models\Owner as Owners;
 use LaravelEnso\ActivityLog\app\Traits\LogActivity;
+use LaravelEnso\Discussions\app\Traits\Discussable;
 use LaravelEnso\CommentsManager\app\Traits\Commentable;
 use LaravelEnso\AddressesManager\app\Traits\Addressable;
 use LaravelEnso\DocumentsManager\app\Traits\Documentable;
 
 class Owner extends Owners
 {
-    use Contactable, Commentable, Documentable, Addressable, LogActivity;
+    use Contactable, Commentable, Discussable, Documentable, Addressable, LogActivity;
 
     protected $fillable = ['name', 'description', 'is_active'];