+++ config/enso/discussions.php
@@ -1,0 +1,14 @@
+<?php
+
+use App\Owner;
+
+return [
+    'discussables' => [
+        'owner' => Owner::class,
+    ],
+    'loggableMorph' => [
+        'discussable' => [
+            Owner::class => 'name',
+        ],
+    ],
+];