+++ resources/js/components/enso/bulma/MorphableContainer.vue
@@ -8,8 +8,7 @@
                 {{ count[tab.toLowerCase()] }}
             </span>
         </span>
-        <tab class="has-padding-medium"
-            :id="__('Addresses')"
+        <tab :id="__('Addresses')"
             v-if="addresses">
             <addresses controls
                 :type="type"
@@ -17,8 +16,7 @@
                 @update="count.addresses = $refs.addresses.count"
                 ref="addresses"/>
         </tab>
-        <tab class="has-padding-medium"
-            :id="__('Comments')"
+        <tab :id="__('Comments')"
             v-if="comments">
             <comments controls
                 :type="type"
@@ -26,8 +24,7 @@
                 @update="count.comments = $refs.comments.count"
                 ref="comments"/>
         </tab>
-        <tab class="has-padding-medium"
-            :id="__('Contacts')"
+        <tab :id="__('Contacts')"
             v-if="contacts">
             <contacts controls
                 :type="type"
@@ -35,8 +32,15 @@
                 @update="count.contacts = $refs.contacts.count"
                 ref="contacts"/>
         </tab>
-        <tab class="has-padding-medium"
-            :id="__('Documents')"
+        <tab :id="__('Discussions')"
+            v-if="discussions">
+            <discussions controls
+                :type="type"
+                :id="id"
+                @update="count.discussions = $refs.discussions.count"
+                ref="discussions"/>
+        </tab>
+        <tab :id="__('Documents')"
             v-if="documents">
             <documents controls
                 :type="type"
@@ -54,13 +58,14 @@ import Tab from './Tab.vue';
 import Addresses from '../addresses/Addresses.vue';
 import Comments from '../comments/Comments.vue';
 import Contacts from '../contacts/Contacts.vue';
+import Discussions from '../discussions/Discussions.vue';
 import Documents from '../documents/Documents.vue';
 
 export default {
     name: 'MorphableContainer',
 
     components: {
-        Tabs, Tab, Addresses, Comments, Contacts, Documents,
+        Tabs, Tab, Addresses, Comments, Contacts, Discussions, Documents,
     },
 
     props: {
@@ -72,6 +77,10 @@ export default {
             type: String,
             required: true,
         },
+        addresses: {
+            type: Boolean,
+            default: false,
+        },
         contacts: {
             type: Boolean,
             default: false,
@@ -80,11 +89,11 @@ export default {
             type: Boolean,
             default: false,
         },
-        documents: {
+        discussions: {
             type: Boolean,
             default: false,
         },
-        addresses: {
+        documents: {
             type: Boolean,
             default: false,
         },
@@ -96,14 +105,11 @@ export default {
                 addresses: 0,
                 comments: 0,
                 contacts: 0,
+                discussions: 0,
                 documents: 0,
             },
         };
     },
-
 };
+
 </script>
-
-<style lang="scss" scoped>
-
-</style>