+++ resources/js/components/enso/contacts/Contacts.vue
@@ -1,5 +1,5 @@
 <template>
-    <div class="has-padding-medium wrapper">
+    <div class="wrapper">
         <div class="controls"
             v-if="controls">
             <button class="button"