+++ resources/js/components/enso/discussions/mention/constants/keys.js
@@ -1,0 +1,9 @@
+const Keys = {
+    TAB: 9,
+    ENTER: 13,
+    ESCAPE: 27,
+    UP: 38,
+    DOWN: 40,
+};
+
+export default Keys;